refactor(Database):split database class
Split CRUD operations into smaller files
